Fetch request now is properly authenticated
Teh request was continuously failing because it wasn't authenticated. This
led to an adventure in the server where we had to first allow the cookie
to be accessed outside of HTTP with httpOnly: false. THEN we had to send
the cookie back with the request, which involves making a custom request
with the cookie header, AND credentials set to 'same-origin'. After that,
it will finally work as expected.